# perf: add async analytics pattern for 82% faster response times