--- a/src/lib/analytics/__tests__/async-analytics.test.ts
+++ b/src/lib/analytics/__tests__/async-analytics.test.ts
@@ -1,0 +1,424 @@
+import { describe, it, expect, beforeEach, vi, afterEach } from 'vitest';
+import { 
+  AsyncAnalytics, 
+  CloudflareAnalytics, 
+  AnalyticsFactory,
+  TrackPerformance,
+  createAnalyticsMiddleware 
+} from '../async-analytics';
+
+describe('AsyncAnalytics', () => {
+  let analytics: AsyncAnalytics;
+  let mockCtx: any;
+  let waitUntilCalls: any[] = [];
+
+  beforeEach(() => {
+    waitUntilCalls = [];
+    mockCtx = {
+      waitUntil: vi.fn((promise: Promise<any>) => {
+        waitUntilCalls.push(promise);
+      }),
+      passThroughOnException: vi.fn(),
+    };
+  });
+
+  afterEach(() => {
+    vi.clearAllMocks();
+  });
+
+  describe('basic tracking', () => {
+    beforeEach(() => {
+      analytics = new AsyncAnalytics(mockCtx, {
+        endpoint: 'https://analytics.test/events',
+        apiKey: 'test-key',
+        batching: false, // Disable batching for simpler tests
+      });
+    });
+
+    it('should track events without blocking', () => {
+      analytics.track('test_event', { foo: 'bar' });
+      
+      // waitUntil should be called immediately
+      expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+      expect(waitUntilCalls).toHaveLength(1);
+    });
+
+    it('should track user events', () => {
+      analytics.trackUser('user123', 'login', { method: 'oauth' });
+      
+      expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+    });
+
+    it('should track page views', () => {
+      analytics.trackPageView('/home', { referrer: '/login' });
+      
+      expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+    });
+
+    it('should track errors', () => {
+      const error = new Error('Test error');
+      analytics.trackError(error, { context: 'test' });
+      
+      expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+    });
+
+    it('should track performance metrics', () => {
+      analytics.trackPerformance('api_call', 150, 'ms');
+      
+      expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe('batching', () => {
+    beforeEach(() => {
+      vi.useFakeTimers();
+      analytics = new AsyncAnalytics(mockCtx, {
+        endpoint: 'https://analytics.test/events',
+        batching: true,
+        batchSize: 3,
+        flushInterval: 1000,
+      });
+    });
+
+    afterEach(() => {
+      vi.useRealTimers();
+    });
+
+    it('should batch events until batch size reached', () => {
+      analytics.track('event1');
+      analytics.track('event2');
+      
+      // No calls yet - still batching
+      expect(mockCtx.waitUntil).not.toHaveBeenCalled();
+      
+      analytics.track('event3');
+      
+      // Batch size reached - should flush
+      expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+    });
+
+    it('should flush after interval', () => {
+      analytics.track('event1');
+      analytics.track('event2');
+      
+      expect(mockCtx.waitUntil).not.toHaveBeenCalled();
+      
+      // Advance time past flush interval
+      vi.advanceTimersByTime(1001);
+      
+      expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+    });
+
+    it('should manually flush pending events', () => {
+      analytics.track('event1');
+      analytics.track('event2');
+      
+      expect(mockCtx.waitUntil).not.toHaveBeenCalled();
+      
+      analytics.flush();
+      
+      expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+    });
+
+    it('should not flush if no events pending', () => {
+      analytics.flush();
+      
+      expect(mockCtx.waitUntil).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('error handling', () => {
+    beforeEach(() => {
+      // Mock fetch to return error response
+      global.fetch = vi.fn().mockResolvedValue({
+        ok: false,
+        status: 500,
+        statusText: 'Internal Server Error',
+      });
+      
+      analytics = new AsyncAnalytics(mockCtx, {
+        endpoint: 'https://analytics.example.com/events',
+        batching: false,
+      });
+    });
+
+    it('should not throw on analytics failure', async () => {
+      const consoleSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
+      
+      analytics.track('test_event');
+      
+      // Wait for the promise to resolve
+      await waitUntilCalls[0];
+      
+      // Should log error but not throw
+      expect(consoleSpy).toHaveBeenCalledWith(
+        expect.stringContaining('Analytics failed: 500')
+      );
+      consoleSpy.mockRestore();
+    });
+  });
+
+  describe('no endpoint configured', () => {
+    beforeEach(() => {
+      analytics = new AsyncAnalytics(mockCtx, {
+        // No endpoint
+        debug: true,
+        batching: false, // Disable batching for immediate send
+      });
+    });
+
+    it('should skip sending when no endpoint', async () => {
+      const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {});
+      
+      analytics.track('test_event');
+      
+      expect(mockCtx.waitUntil).toHaveBeenCalled();
+      
+      // Wait for the promise to complete
+      if (waitUntilCalls.length > 0) {
+        await waitUntilCalls[0];
+      }
+      
+      expect(consoleSpy).toHaveBeenCalledWith(
+        expect.stringContaining('No endpoint configured')
+      );
+      
+      consoleSpy.mockRestore();
+    });
+  });
+});
+
+describe('CloudflareAnalytics', () => {
+  let analytics: CloudflareAnalytics;
+  let mockCtx: any;
+  let mockAnalyticsEngine: any;
+
+  beforeEach(() => {
+    mockCtx = {
+      waitUntil: vi.fn(),
+      passThroughOnException: vi.fn(),
+    };
+    
+    mockAnalyticsEngine = {
+      writeDataPoint: vi.fn(),
+    };
+    
+    analytics = new CloudflareAnalytics(mockCtx, mockAnalyticsEngine, {
+      batching: false,
+    });
+  });
+
+  it('should write to Analytics Engine', () => {
+    analytics.track('test_event', { value: 42 });
+    
+    // Since sendToAnalytics is private, we can't directly test it
+    // But we can verify waitUntil was called
+    expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+  });
+});
+
+describe('AnalyticsFactory', () => {
+  it('should create CloudflareAnalytics when Analytics Engine available', () => {
+    const mockCtx = {
+      waitUntil: vi.fn(),
+      passThroughOnException: vi.fn(),
+    };
+    
+    const env = {
+      ANALYTICS_ENGINE: { writeDataPoint: vi.fn() },
+    };
+    
+    const analytics = AnalyticsFactory.create(mockCtx, env);
+    
+    expect(analytics).toBeInstanceOf(CloudflareAnalytics);
+  });
+
+  it('should create AsyncAnalytics when no Analytics Engine', () => {
+    const mockCtx = {
+      waitUntil: vi.fn(),
+      passThroughOnException: vi.fn(),
+    };
+    
+    const env = {
+      ANALYTICS_ENDPOINT: 'https://analytics.test',
+      ANALYTICS_API_KEY: 'test-key',
+    };
+    
+    const analytics = AnalyticsFactory.create(mockCtx, env);
+    
+    expect(analytics).toBeInstanceOf(AsyncAnalytics);
+  });
+
+  it('should create no-op analytics', () => {
+    const analytics = AnalyticsFactory.createNoop();
+    
+    expect(analytics).toBeInstanceOf(AsyncAnalytics);
+    
+    // Should not throw
+    analytics.track('test');
+  });
+});
+
+describe('@TrackPerformance decorator', () => {
+  // Skip decorator tests due to TypeScript config requirements
+  it.skip('should track method performance', async () => {
+    expect(true).toBe(true);
+  });
+});
+
+describe('createAnalyticsMiddleware', () => {
+  it('should track successful requests', async () => {
+    const mockAnalytics = {
+      track: vi.fn(),
+    };
+    
+    const middleware = createAnalyticsMiddleware(() => mockAnalytics as any);
+    
+    const ctx = {
+      request: {
+        url: '/test',
+        method: 'GET',
+      },
+      response: {
+        status: 200,
+      },
+    };
+    
+    const next = vi.fn().mockResolvedValue(undefined);
+    
+    await middleware(ctx, next);
+    
+    expect(next).toHaveBeenCalled();
+    expect(mockAnalytics.track).toHaveBeenCalledWith(
+      'request_completed',
+      expect.objectContaining({
+        path: '/test',
+        method: 'GET',
+        status: 200,
+      })
+    );
+  });
+
+  it('should track request errors', async () => {
+    const mockAnalytics = {
+      track: vi.fn(),
+    };
+    
+    const middleware = createAnalyticsMiddleware(() => mockAnalytics as any);
+    
+    const ctx = {
+      request: {
+        url: '/test',
+        method: 'POST',
+      },
+    };
+    
+    const error = new Error('Request failed');
+    const next = vi.fn().mockRejectedValue(error);
+    
+    await expect(middleware(ctx, next)).rejects.toThrow('Request failed');
+    
+    expect(mockAnalytics.track).toHaveBeenCalledWith(
+      'request_error',
+      expect.objectContaining({
+        path: '/test',
+        method: 'POST',
+        error: 'Request failed',
+      })
+    );
+  });
+});
+
+describe('Production Scenarios', () => {
+  let analytics: AsyncAnalytics;
+  let mockCtx: any;
+
+  beforeEach(() => {
+    mockCtx = {
+      waitUntil: vi.fn(),
+      passThroughOnException: vi.fn(),
+    };
+    
+    global.fetch = vi.fn().mockResolvedValue({
+      ok: true,
+      status: 200,
+    });
+  });
+
+  it('should handle high-volume event tracking', () => {
+    analytics = new AsyncAnalytics(mockCtx, {
+      endpoint: 'https://analytics.test',
+      batching: true,
+      batchSize: 100,
+    });
+    
+    // Simulate high volume
+    for (let i = 0; i < 1000; i++) {
+      analytics.track(`event_${i}`, { index: i });
+    }
+    
+    // Should batch efficiently
+    // 1000 events / 100 batch size = 10 batches
+    expect(mockCtx.waitUntil).toHaveBeenCalledTimes(10);
+  });
+
+  it('should not block response time', () => {
+    analytics = new AsyncAnalytics(mockCtx, {
+      endpoint: 'https://analytics.test',
+      batching: false,
+    });
+    
+    const start = Date.now();
+    
+    // Track 100 events
+    for (let i = 0; i < 100; i++) {
+      analytics.track('event', { i });
+    }
+    
+    const duration = Date.now() - start;
+    
+    // Should complete almost instantly (< 10ms)
+    // Events are sent async via waitUntil
+    expect(duration).toBeLessThan(10);
+    expect(mockCtx.waitUntil).toHaveBeenCalledTimes(100);
+  });
+
+  it('should track Telegram bot update lifecycle', () => {
+    analytics = new AsyncAnalytics(mockCtx, {
+      endpoint: 'https://analytics.test',
+      batching: true,
+      batchSize: 10,
+    });
+    
+    // Simulate bot update handling
+    const updateId = '12345';
+    const userId = 'user123';
+    
+    // Track update received
+    analytics.track('telegram_update_received', {
+      update_id: updateId,
+      type: 'message',
+    });
+    
+    // Track user action
+    analytics.trackUser(userId, 'command_executed', {
+      command: '/start',
+      update_id: updateId,
+    });
+    
+    // Track processing time
+    analytics.trackPerformance('update_processing', 45);
+    
+    // Track response sent
+    analytics.track('telegram_response_sent', {
+      update_id: updateId,
+      response_type: 'text',
+    });
+    
+    // Flush at end of request
+    analytics.flush();
+    
+    // Should batch all events into single call
+    expect(mockCtx.waitUntil).toHaveBeenCalledTimes(1);
+  });
+});