--- a/src/lib/analytics/index.ts
+++ b/src/lib/analytics/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Analytics Module
+ * 
+ * Fire-and-forget analytics for Cloudflare Workers
+ * 
+ * @module lib/analytics
+ */
+
+export {
+  AsyncAnalytics,
+  CloudflareAnalytics,
+  AnalyticsFactory,
+  TrackPerformance,
+  createAnalyticsMiddleware,
+  type AnalyticsEvent,
+  type AsyncAnalyticsOptions,
+} from './async-analytics';