… docs/patterns/async-analytics.md
… src/lib/analytics/examples/telegram-bot-analytics.ts